Added HTML
Cannot seem to get some of the tests to work. Specifically the "User adding a tour to their wishlist". I had no issue recording the actions, but every time i tested it would fail.

The "Unsuccessful supplier login" test was timing out when ran with the test suite which also confused me because it ran just fine on its own.